+++ lib/philomena/users/user.ex
@@ -250,8 +250,9 @@ defmodule Philomena.Users.User do
     |> cast(attrs, [:scratchpad])
   end
 
-  def username_changeset(user, attrs) do
+  def name_changeset(user, attrs) do
     now = NaiveDateTime.utc_now() |> NaiveDateTime.truncate(:second)
+
     user
     |> cast(attrs, [:name])
     |> validate_required([:name])
