+++ lib/philomena_web/views/registration/name_view.ex
@@ -1,0 +1,3 @@
+defmodule PhilomenaWeb.Registration.NameView do
+  use PhilomenaWeb, :view
+end
