--- lib/philomena_web/views/registration/username_view.ex
@@ -1,3 +1,0 @@
-defmodule PhilomenaWeb.Registration.UsernameView do
-    use PhilomenaWeb, :view
-end
